[added] Full cart & orders interactions (CRUD); Policies for models (security & roles); Small additions; [deleted] Refactored routes; Some little changes in Middlewares;

+++ app/Http/Controllers/ReviewController.php
@@ -6,10 +6,20 @@ use App\Http\Resources\ReviewResource;
 use App\Models\Review;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller as BaseController;
+use Illuminate\Support\Facades\Gate;
 
 
 class ReviewController extends BaseController
 {
+    public function reviewsForBike(Request $request,int $bike_id)
+    {
+        $reviews = Review::where('bike_id', $bike_id)->paginate(5);
+        if (!$reviews) {
+            return response()->json(['message' => 'There are no reviews for this bike!'], 404);
+        }
+        return new ReviewCollection($reviews);
+    }
+
     public function index()
     {
         $reviews = Review::paginate(5);
@@ -29,8 +39,11 @@ class ReviewController extends BaseController
     public function store(Request $request)
     {
         $review = new Review;
+        if ($request->user()->cannot('create', $review)) {
+            abort(403, 'Unauthorized');
+        }
         $review->bike_id = $request->bike_id;
-        $review->user_id = $request->user_id;
+        $review->user_id = $request->user()->id;
         $review->text = $request->text;
         $review->rating = $request->rating;
         $review->save();
@@ -40,10 +53,12 @@ class ReviewController extends BaseController
 
     public function update(Request $request, int $id)
     {
+
         $review = Review::find($id);
+        if ($request->user()->cannot('update', $review)) {
+            abort(403, 'Unauthorized');
+        }
         if ($review) {
-            $review->bike_id = $request->bike_id;
-            $review->user_id = $request->user_id;
             $review->text = $request->text;
             $review->rating = $request->rating;
             $review->save();
@@ -53,9 +68,12 @@ class ReviewController extends BaseController
         }
     }
 
-    public function destroy(int $id)
+    public function destroy(Request $request, int $id)
     {
         $review = Review::find($id);
+        if ($request->user()->cannot('delete', $review)) {
+            abort(403, 'Unauthorized');
+        }
         if ($review) {
             $review->delete();
             return response()->json(['message' => 'Review deleted']);
